add test python warp c so

+++ PythonWarpTest/hello.c
@@ -2,16 +2,15 @@
 
 #include <stdio.h>
 #include <stdlib.h>
-#include "testheader.h"
 
 
 int foo(int a, int b)  
-{  
+{
   printf("you input %d and %d\n", a, b);  
   return a+b;  
 }
 
-double get_double(int a, int b, double c)
+int get_double(int a, int b, int c)
 {
 	int acount = a + b;
 	if(acount > c)
